requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.5.0
motor==3.3.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
modal==0.73.72
pytube==15.0.0
python-jose==3.3.0
google-api-python-client==2.115.0